Preparse: Report UTF8 errors 1st, then ctrl chars
Our preparse routines will always return an error in the presence of
invalid UTF8 and/or ASCII control characters, but previously they might
have reported some other parsing error instead of these foundational
errors.

Now if there's a UTF8 error we report that, and for valid UTF8 strings
we report the presence of control characters before any other parsing
error.

And we're almost back to counting the results of preparse and
regex_preparse as identical if they both error — except that if either
returns a UTF8 or control-character error, the other must return the
same error or we consider the results different.